Seperate component for draw stuff
